+++ entry/main.go
@@ -3,7 +3,7 @@ package main
 import "fmt"
 
 func init() {
-		
+
 }
 
 func main() {
