+++ routes/payloads/scheme.go
@@ -1,6 +1,6 @@
 package payloads
 
-import "restbomber/models"
+import "github.com/bomber-team/rest-bomber/models"
 
 /*SchemePayload - payload for configuring datas models*/
 type SchemePayload struct {
