--- tests/tests_data/scenario.json
@@ -1,14 +1,0 @@
-{
-  "headers": {
-    "method": "GET",
-    "authority": "bla"
-  },
-  "body": {
-    "Name": "WordGenerator",
-    "Password": "PasswordGenerator",
-    "Nas-Ip": "IpGeneratorV6",
-    "test": {
-      "test": "WordGenerator"
-    }
-  }
-}
